Refactor to break code out of lib.rs into separate modules and place test code into a test directory.
