--- src/infallible/mod.rs
@@ -1,3 +1,0 @@
-mod rwlock;
-
-pub use rwlock::{RwLock, RwLockReadGuard, RwLockWriteGuard};
