+++ src/reader.rs
@@ -1,0 +1,20 @@
+use crate::node_type::{LeafNode, Node, NodeKey};
+use anyhow::{format_err, Result};
+
+/// `TreeReader` defines the interface between
+/// [`JellyfishMerkleTree`](struct.JellyfishMerkleTree.html)
+/// and underlying storage holding nodes.
+pub trait TreeReader<V> {
+    /// Gets node given a node key. Returns error if the node does not exist.
+    fn get_node(&self, node_key: &NodeKey) -> Result<Node<V>> {
+        self.get_node_option(node_key)?
+            .ok_or_else(|| format_err!("Missing node at {:?}.", node_key))
+    }
+
+    /// Gets node given a node key. Returns `None` if the node does not exist.
+    fn get_node_option(&self, node_key: &NodeKey) -> Result<Option<Node<V>>>;
+
+    /// Gets the rightmost leaf. Note that this assumes we are in the process of restoring the tree
+    /// and all nodes are at the same version.
+    fn get_rightmost_leaf(&self) -> Result<Option<(NodeKey, LeafNode<V>)>>;
+}
