+++ src/tests/nibble_path.rs
@@ -1,8 +1,10 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use super::{arb_internal_nibble_path, skip_common_prefix, NibblePath};
-use crate::types::nibble::Nibble;
+use crate::types::nibble::{
+    nibble_path::{arb_internal_nibble_path, skip_common_prefix, NibblePath},
+    Nibble,
+};
 use proptest::prelude::*;
 
 #[test]
