+++ src/types/nibble.rs
@@ -1,0 +1,145 @@
+// Copyright (c) The Diem Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+#![forbid(unsafe_code)]
+
+//! `Nibble` represents a four-bit unsigned integer.
+
+pub mod nibble_path;
+
+use crate::hash::HashValue;
+#[cfg(any(test, feature = "fuzzing"))]
+use proptest::prelude::*;
+use serde::{Deserialize, Serialize};
+use std::fmt;
+
+/// The hardcoded maximum height of a state merkle tree in nibbles.
+pub const ROOT_NIBBLE_HEIGHT: usize = HashValue::LENGTH * 2;
+
+#[derive(Clone, Copy, Debug, Hash, Eq, PartialEq, Ord, PartialOrd, Serialize, Deserialize)]
+pub struct Nibble(u8);
+
+impl From<u8> for Nibble {
+    fn from(nibble: u8) -> Self {
+        assert!(nibble < 16, "Nibble out of range: {}", nibble);
+        Self(nibble)
+    }
+}
+
+impl From<Nibble> for u8 {
+    fn from(nibble: Nibble) -> Self {
+        nibble.0
+    }
+}
+
+impl fmt::LowerHex for Nibble {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{:x}", self.0)
+    }
+}
+
+/// An iterator that iterates the index range (inclusive) of each different nibble at given
+/// `nibble_idx` of all the keys in a sorted key-value pairs.
+pub(crate) struct NibbleRangeIterator<'a, V> {
+    sorted_kvs: &'a [(HashValue, V)],
+    nibble_idx: usize,
+    pos: usize,
+}
+
+impl<'a, V> NibbleRangeIterator<'a, V> {
+    pub fn new(sorted_kvs: &'a [(HashValue, V)], nibble_idx: usize) -> Self {
+        assert!(nibble_idx < ROOT_NIBBLE_HEIGHT);
+        NibbleRangeIterator {
+            sorted_kvs,
+            nibble_idx,
+            pos: 0,
+        }
+    }
+}
+
+impl<'a, V> std::iter::Iterator for NibbleRangeIterator<'a, V> {
+    type Item = (usize, usize);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let left = self.pos;
+        if self.pos < self.sorted_kvs.len() {
+            let cur_nibble: u8 = self.sorted_kvs[left].0.nibble(self.nibble_idx);
+            let (mut i, mut j) = (left, self.sorted_kvs.len() - 1);
+            // Find the last index of the cur_nibble.
+            while i < j {
+                let mid = j - (j - i) / 2;
+                if self.sorted_kvs[mid].0.nibble(self.nibble_idx) > cur_nibble {
+                    j = mid - 1;
+                } else {
+                    i = mid;
+                }
+            }
+            self.pos = i + 1;
+            Some((left, i))
+        } else {
+            None
+        }
+    }
+}
+
+#[cfg(any(test, feature = "fuzzing"))]
+impl Arbitrary for Nibble {
+    type Parameters = ();
+    type Strategy = BoxedStrategy<Self>;
+
+    fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
+        (0..16u8).prop_map(Self::from).boxed()
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use crate::hash::TestOnlyHash;
+
+    use super::*;
+    #[test]
+    fn test_common_prefix_nibbles_len() {
+        {
+            let hash1 = b"hello".test_only_hash();
+            let hash2 = b"HELLO".test_only_hash();
+            assert_eq!(hash1[0], 0b0011_0011);
+            assert_eq!(hash2[0], 0b1011_1000);
+            assert_eq!(hash1.common_prefix_nibbles_len(hash2), 0);
+        }
+        {
+            let hash1 = b"hello".test_only_hash();
+            let hash2 = b"world".test_only_hash();
+            assert_eq!(hash1[0], 0b0011_0011);
+            assert_eq!(hash2[0], 0b0100_0010);
+            assert_eq!(hash1.common_prefix_nibbles_len(hash2), 0);
+        }
+        {
+            let hash1 = b"hello".test_only_hash();
+            let hash2 = b"100011001000".test_only_hash();
+            assert_eq!(hash1[0], 0b0011_0011);
+            assert_eq!(hash2[0], 0b0011_0011);
+            assert_eq!(hash1[1], 0b0011_1000);
+            assert_eq!(hash2[1], 0b0010_0010);
+            assert_eq!(hash1.common_prefix_nibbles_len(hash2), 2);
+        }
+        {
+            let hash1 = b"hello".test_only_hash();
+            let hash2 = b"hello".test_only_hash();
+            assert_eq!(
+                hash1.common_prefix_nibbles_len(hash2),
+                HashValue::LENGTH * 2
+            );
+        }
+    }
+
+    #[test]
+    fn test_get_nibble() {
+        let hash = b"hello".test_only_hash();
+        assert_eq!(hash.get_nibble(0), Nibble::from(3));
+        assert_eq!(hash.get_nibble(1), Nibble::from(3));
+        assert_eq!(hash.get_nibble(2), Nibble::from(3));
+        assert_eq!(hash.get_nibble(3), Nibble::from(8));
+        assert_eq!(hash.get_nibble(62), Nibble::from(9));
+        assert_eq!(hash.get_nibble(63), Nibble::from(2));
+    }
+}
